requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
appengine-python-standard==1.1.6
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docstring_parser==0.16
Flask==3.0.3
Flask-Dropzone==2.0.0
Flask-Session==0.8.0
flatbuffers==24.3.25
fonttools==4.50.0
frozendict==2.4.0
gast==0.4.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.47.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-firestore==2.16.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
hashring==1.5.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.14.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
memcache==0.12.0
mistune==3.0.2
ml-dtypes==0.2.0
mock==5.1.0
msgspec==0.18.6
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.5.0
Pygments==2.17.2
pymemcache==4.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shapely==2.0.3
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.2
wrapt==1.14.1
